refactor(ast): split AstFolder out of ast.rs into ast_folder.rs
`ast.rs` was already 3800+ lines of AST type definitions; tacking the
fold infrastructure on the bottom only made it harder to navigate. Move
the `AstFolder` trait, all `default_fold_*` helpers, and the
`NoopFolder` round-trip test into a new sibling module
`ast_folder.rs` (matching the existing `ast_index.rs` split).

Importers (`desugar.rs`, `resolver/assert.rs`) now pull `AstFolder` /
`default_fold_*` from `crate::ast_folder` explicitly, which also
documents the dependency direction: the folder builds on top of the AST
types, not part of them.

`ast.rs` drops from 3824 → 3179 lines; no behavior change.
